Display status of task completion
Use a reducer to define how different actions (set robot, start a task,
and finish a task) should change the state. Without the reducer, it is
hard to set working, currentTask, etc. for the robot at the same time.

As a side effect of rendering with a currentTask, the finish task action
is scheduled, an async function is used to await the robot to complete
the task based on currentTask's duration.